% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeb.R
\name{bf2se}
\alias{bf2se}
\title{SE for BF}
\usage{
bf2se(mcrun, transf = c("no", "mu", "wo"))
}
\arguments{
\item{mcrun}{The output from the function \code{mcsglmm}.}

\item{transf}{The type of transformation to use.}
}
\value{
The precision matrix
}
\description{
Standard errors for BF estimates
}
\details{
Using the formula from Casella.
}
