% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{contour.bfsp}
\alias{contour.bfsp}
\title{Plot the estimated Bayes factors}
\usage{
\method{contour}{bfsp}(x, ipar, xlab, ylab, main, ...)
}
\arguments{
\item{x}{Output from the function \code{\link{bf2new}}.}

\item{ipar}{One or two unique integers in 1:4 indicating with
respect to which parameters to plot.}

\item{xlab,ylab,main}{Graphical parameters}

\item{...}{Other graphical parameters to be passed on either
\code{plot} or \code{contour}.}
}
\value{
Produces the plot mentioned in the Details.
}
\description{
Creates a plot of the estimated Bayes factors from the function
\code{\link{bf2new}}.
}
\details{
Depending on whether \code{ipar} is of length 1 or 2, this
function creates a line or a contour plot of the estimated Bayes
factors.
}

