% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{package}
\name{geoBayes}
\alias{geoBayes}
\alias{geoBayes-package}
\title{The \code{geoBayes} package}
\description{
geoBayes
}
\details{
Analysis of geostatistical data using Bayes and Empirical Bayes
methods.

This package provides functions to fit geostatistical data. The
data can be continuous, binary or count data and the models
implemented are flexible. Conjugate priors are assumed on some
parameters while non-conjugate parameters can be estimated either
by Monte-Carlo or empirical Bayes methods.
}
\author{
Evangelos Evangelou <e.evangelou@maths.bath.ac.uk> and Vivekanda
Roy <vroy@iastate.edu>
}
\references{
Roy, V., Evangelou, E. and Zhu, Z. (2015). Empirical
Bayes methods for the transformed Gaussian random fields model
with additive measurement errors. In Upadhyay, S. K., Singh, U.,
Dey, D. K., and Loganathan, A., editors, \emph{Current Trends in
Bayesian Methodology with Applications}, Boca Raton, FL, USA, CRC
Press.
}
\seealso{
\code{geoR}, \code{geoRglm}
}
\keyword{package}

