% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobayes-correlation.R
\docType{data}
\name{geoBayes_correlation}
\alias{geoBayes_correlation}
\alias{.geoBayes_corrfcn}
\title{Spatial correlation used in the geoBayes package}
\format{
An object of class \code{data.frame} with 5 rows and 4 columns.
}
\usage{
.geoBayes_corrfcn
}
\description{
This hidden variable contains a choice of correlation functions
that can be fit with this package. The function can be chosen in
the \code{corrfcn} input of the relevant function. This variable
cannot be overwritten.
}
\keyword{datasets}
