% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobayes-models.R
\docType{data}
\name{geoBayes_models}
\alias{geoBayes_models}
\alias{.geoBayes_models}
\title{Models used in the geoBayes package}
\format{An object of class \code{data.frame} with 12 rows and 7 columns.}
\usage{
.geoBayes_models
}
\description{
This hidden variable contains a choice of models that can be fit
with this package. The model can be chosen in the \code{family}
input of the relevant function. This variable cannot be
overwritten.
}
\keyword{datasets}
