% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{mcmcmake}
\alias{mcmcmake}
\title{Convert to an \code{\link[coda]{mcmc}} object}
\usage{
mcmcmake(...)
}
\arguments{
\item{...}{Output(s) from the functions mentioned in the Details.}
}
\value{
An mcmc object.
}
\description{
Convert to an \code{\link[coda]{mcmc}} object.
}
\details{
This function takes as input the one or more output(s) from
function \code{\link{mcsglmm}} or \code{\link{mcstrga}} and
returns an \code{\link[coda]{mcmc}} object or an
\code{\link[coda]{mcmc.list}} object for coda. The function
requires the \code{coda} package to be installed.
}

