% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{robitinv}
\alias{robitinv}
\title{Compute the inverse logit, probit, robit transformations}
\usage{
robitinv(x, d, log.p = TRUE)
}
\arguments{
\item{x}{Numerical input at which the inverse link function is to
be computed.}

\item{d}{Defines the link function. See details.}

\item{log.p}{Whether to return the \code{log} of the inverse link.}
}
\value{
A numerical output of the same lenght and dimension as
\code{x}.
}
\description{
Compute the inverse logit, probit, robit transformations.
}
\details{
The input argument \code{d} determines the link function. If
\code{"logit"} or \code{"probit"}, then the respective link
function is used. If a positive integer, then the robit(\code{d}),
i.e. the CDF of the Student's t with degrees of freedom \code{d}
is used. Note that, contrary to the common, the inverse logit
transformation is defined here as the CDF of the logistic
distribution with scale \code{0.6458} to make it comparable with
the robit link.
}

