% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sploglik.R
\name{sploglik}
\alias{sploglik}
\title{Spatial log likelihood}
\usage{
sploglik(pargrid, runs, transf = c("no", "mu", "wo"))
}
\arguments{
\item{pargrid}{A data frame with components "linkp", "phi", "omg",
"kappa". Each row gives a combination of the parameters to compute
the log-likelihood.}

\item{runs}{A list with outputs from the function
\code{\link{mcsglmm}} or \code{\link{mcstrga}}.}

\item{transf}{Whether to use a transformed sample for the
computations. If \code{"no"} or \code{FALSE}, it doesn't. If
\code{"mu"} or \code{TRUE}, it uses the samples for the mean. If
\code{"wo"} it uses an alternative transformation. The latter
can be used only for the families indicated by
\code{.geoBayes_models$haswo}.}
}
\value{
A matrix with number of rows the total number of samples
  in runs and number of columns the number of rows in pargrid. The
  [i,j] element of the matrix is the value of the loglikelihood at
  the ith sample when all samples in runs are put together evaluated
  at the jth parameter value.
}
\description{
Spatial joint log likelihood
}
\details{
Computes the joint log likelihood log f(y,z|parameters) or log
f(y,mu|parameters) for each (y,z) or (y,mu) in runs and for parameters
  in pargrid up to a constant which does not depend on the parameters.
}
