% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcsp2.R
\name{subset.geomcmc}
\alias{subset.geomcmc}
\title{Subset MCMC chain}
\usage{
\method{subset}{geomcmc}(x, subset, ...)
}
\arguments{
\item{x}{Output from the functions \code{\link{mcsglmm}} or
\code{\link{mcstrga}}.}

\item{subset}{Logical or integer vector.}

\item{...}{Further arguments to be passed to or from other methods.}
}
\value{
A similar object as \code{x} with the subsetted chain.
}
\description{
Return subset of MCMC chain.
}
