% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{transfsample}
\alias{transfsample}
\title{Compute transformed sample}
\usage{
transfsample(runs, model, transf = c("no", "mu", "wo"))
}
\arguments{
\item{runs}{A list of lists with elements z, mu, nu, whichobs
which gives the samples and link parameter.}

\item{model}{A list with elements response and family.}

\item{transf}{The type of transformation to use.}
}
\value{
A list with elements sample, transf, itr, ifam.
}
\description{
Compute transformed sample.
}
\details{
Computes the transformed sample according to transf.
}
